refactor other functionalities from ast parser into separate classes
