+++ sql_metadata/dialect_parser.py
@@ -1,0 +1,192 @@
+"""SQL dialect detection, parsing, and parse-quality validation.
+
+Combines dialect heuristics (which sqlglot dialect to try), the actual
+``sqlglot.parse()`` call, and degraded-result detection into a single
+class so that callers only need to call :meth:`DialectParser.parse`.
+"""
+
+import logging
+
+import sqlglot
+from sqlglot import Dialect, exp
+from sqlglot.dialects.tsql import TSQL
+from sqlglot.errors import ParseError, TokenError
+from sqlglot.tokens import Tokenizer
+
+from sql_metadata.comments import _has_hash_variables
+
+#: Table names that indicate a degraded parse result.
+_BAD_TABLE_NAMES = frozenset({"IGNORE", ""})
+
+#: SQL keywords that should not appear as bare column names.
+_BAD_COLUMN_NAMES = frozenset({"UNIQUE", "DISTINCT", "SELECT", "FROM", "WHERE"})
+
+
+# ---------------------------------------------------------------------------
+# Custom dialect classes
+# ---------------------------------------------------------------------------
+
+
+class HashVarDialect(Dialect):
+    """Custom sqlglot dialect that treats ``#WORD`` as identifiers.
+
+    MSSQL uses ``#`` to prefix temporary table names (e.g. ``#temp``)
+    and some template engines use ``#VAR#`` placeholders.  The default
+    sqlglot tokenizer treats ``#`` as an unknown single-character token;
+    this dialect moves it into ``VAR_SINGLE_TOKENS`` so it becomes part
+    of a ``VAR`` token instead.
+    """
+
+    class Tokenizer(Tokenizer):
+        """Tokenizer subclass that includes ``#`` in variable tokens."""
+
+        SINGLE_TOKENS = {**Tokenizer.SINGLE_TOKENS}
+        SINGLE_TOKENS.pop("#", None)
+        VAR_SINGLE_TOKENS = {*Tokenizer.VAR_SINGLE_TOKENS, "#"}
+
+
+class BracketedTableDialect(TSQL):
+    """TSQL dialect for queries containing ``[bracketed]`` identifiers.
+
+    sqlglot's TSQL dialect correctly interprets square-bracket quoting,
+    which the default dialect does not.  This thin subclass exists so
+    that ``TableExtractor`` can ``isinstance``-check to enable
+    bracket-preserving table name construction.
+    """
+
+
+# ---------------------------------------------------------------------------
+# DialectParser
+# ---------------------------------------------------------------------------
+
+
+class DialectParser:
+    """Detect the appropriate sqlglot dialect and parse SQL into an AST."""
+
+    def parse(self, clean_sql: str) -> tuple:
+        """Parse *clean_sql*, returning ``(ast, dialect)``.
+
+        Detects candidate dialects via heuristics, tries each in order,
+        and returns the first non-degraded result.
+
+        :param clean_sql: Preprocessed SQL string (comments stripped, etc.).
+        :type clean_sql: str
+        :returns: 2-tuple of ``(ast_node, winning_dialect)``.
+        :rtype: tuple
+        :raises ValueError: If all dialect attempts fail.
+        """
+        dialects = self._detect_dialects(clean_sql)
+        return self._try_dialects(clean_sql, dialects)
+
+    # -- dialect detection --------------------------------------------------
+
+    @staticmethod
+    def _detect_dialects(sql: str) -> list:
+        """Choose an ordered list of sqlglot dialects to try for *sql*.
+
+        Heuristics:
+
+        * ``#WORD`` → :class:`HashVarDialect` (MSSQL temp tables).
+        * Back-ticks → ``"mysql"``.
+        * Square brackets or ``TOP`` → :class:`BracketedTableDialect`.
+        * ``UNIQUE`` → try default, MySQL, Oracle.
+        * ``LATERAL VIEW`` → ``"spark"`` (Hive).
+
+        :param sql: Cleaned SQL string.
+        :type sql: str
+        :returns: Ordered list of dialects to attempt.
+        :rtype: list
+        """
+        upper = sql.upper()
+        if _has_hash_variables(sql):
+            return [HashVarDialect, None, "mysql"]
+        if "`" in sql:
+            return ["mysql", None]
+        if "[" in sql or " TOP " in upper:
+            return [BracketedTableDialect, None, "mysql"]
+        if " UNIQUE " in upper:
+            return [None, "mysql", "oracle"]
+        if "LATERAL VIEW" in upper:
+            return ["spark", None, "mysql"]
+        return [None, "mysql"]
+
+    # -- parsing ------------------------------------------------------------
+
+    def _try_dialects(self, clean_sql: str, dialects: list) -> tuple:
+        """Try parsing *clean_sql* with each dialect, returning the best.
+
+        :returns: 2-tuple of ``(ast_node, winning_dialect)``.
+        :raises ValueError: If all dialect attempts fail.
+        """
+        last_result = None
+        winning_dialect = None
+        for dialect in dialects:
+            try:
+                result = self._parse_with_dialect(clean_sql, dialect)
+                if result is None:
+                    continue
+                last_result = result
+                winning_dialect = dialect
+                is_last = dialect == dialects[-1]
+                if not is_last and self._is_degraded(result, clean_sql):
+                    continue
+                return result, dialect
+            except (ParseError, TokenError):
+                if dialect is not None and dialect == dialects[-1]:
+                    raise ValueError("This query is wrong")
+                continue
+
+        if last_result is not None:
+            return last_result, winning_dialect
+        raise ValueError("This query is wrong")
+
+    @staticmethod
+    def _parse_with_dialect(clean_sql: str, dialect) -> exp.Expression:
+        """Parse *clean_sql* with a single dialect, suppressing warnings."""
+        logger = logging.getLogger("sqlglot")
+        old_level = logger.level
+        logger.setLevel(logging.CRITICAL)
+        try:
+            results = sqlglot.parse(
+                clean_sql,
+                dialect=dialect,
+                error_level=sqlglot.ErrorLevel.WARN,
+            )
+        finally:
+            logger.setLevel(old_level)
+
+        if not results or results[0] is None:
+            return None
+        result = results[0]
+        if isinstance(result, exp.Subquery) and not result.alias:
+            result = result.this
+        return result
+
+    # -- quality checks -----------------------------------------------------
+
+    def _is_degraded(self, result: exp.Expression, clean_sql: str) -> bool:
+        """Return ``True`` when a better dialect should be tried."""
+        if isinstance(result, exp.Command) and not self._is_expected_command(clean_sql):
+            return True
+        return self._has_parse_issues(result)
+
+    @staticmethod
+    def _is_expected_command(sql: str) -> bool:
+        """Check whether *sql* legitimately parses as ``exp.Command``."""
+        upper = sql.strip().upper()
+        return upper.startswith("CREATE FUNCTION")
+
+    @staticmethod
+    def _has_parse_issues(ast: exp.Expression) -> bool:
+        """Detect signs of a degraded or incorrect parse.
+
+        Checks for table nodes with empty/keyword-like names and column
+        nodes whose name is a SQL keyword without a table qualifier.
+        """
+        for table in ast.find_all(exp.Table):
+            if table.name in _BAD_TABLE_NAMES:
+                return True
+        for col in ast.find_all(exp.Column):
+            if col.name.upper() in _BAD_COLUMN_NAMES and not col.table:
+                return True
+        return False
