+++ sql_metadata/sql_cleaner.py
@@ -1,0 +1,176 @@
+"""Raw SQL preprocessing before AST construction.
+
+Pure string transformations — no sqlglot dependency.  Handles comment
+stripping, ``REPLACE INTO`` rewriting, qualified CTE name normalisation,
+DB2 isolation-level clauses, malformed-query rejection, and redundant
+outer-parenthesis removal.
+"""
+
+import itertools
+import re
+from typing import NamedTuple, Optional
+
+from sql_metadata.comments import strip_comments_for_parsing as _strip_comments
+
+
+class CleanResult(NamedTuple):
+    """Result of :meth:`SqlCleaner.clean`."""
+
+    sql: Optional[str]
+    is_replace: bool
+    cte_name_map: dict
+
+
+def _strip_outer_parens(sql: str) -> str:
+    """Strip redundant outer parentheses from *sql*.
+
+    Needed because sqlglot cannot parse double-wrapped non-SELECT
+    statements like ``((UPDATE ...))``.  Uses ``itertools.accumulate``
+    to verify balanced parens in one pass, with recursion for nesting.
+    """
+    s = sql.strip()
+
+    def _is_wrapped(text):
+        if len(text) < 2 or text[0] != "(" or text[-1] != ")":
+            return False
+        inner = text[1:-1]
+        depths = list(
+            itertools.accumulate(
+                (1 if c == "(" else -1 if c == ")" else 0) for c in inner
+            )
+        )
+        return not depths or min(depths) >= 0
+
+    if _is_wrapped(s):
+        return _strip_outer_parens(s[1:-1].strip())
+    return s
+
+
+def _normalize_cte_names(sql: str) -> tuple:
+    """Replace qualified CTE names with simple placeholders.
+
+    sqlglot cannot parse ``WITH db.cte_name AS (...)`` because it
+    interprets ``db.cte_name`` as a table reference.  This function
+    rewrites such names to ``db__DOT__cte_name`` and returns a mapping
+    so that the original qualified names can be restored after extraction.
+
+    :param sql: SQL string that may contain qualified CTE names.
+    :type sql: str
+    :returns: A 2-tuple of ``(modified_sql, {placeholder: original_name})``.
+    :rtype: tuple
+    """
+    name_map = {}
+    # Find WITH ... AS patterns with qualified names
+    pattern = re.compile(
+        r"(\bWITH\s+|,\s*)(\w+\.\w+)(\s+AS\s*\()",
+        re.IGNORECASE,
+    )
+
+    def replacer(match):
+        prefix = match.group(1)
+        qualified_name = match.group(2)
+        suffix = match.group(3)
+        placeholder = qualified_name.replace(".", "__DOT__")
+        name_map[placeholder] = qualified_name
+        return f"{prefix}{placeholder}{suffix}"
+
+    modified = pattern.sub(replacer, sql)
+
+    # Also replace references to qualified CTE names in FROM/JOIN clauses
+    for placeholder, original in name_map.items():
+        # Replace references but not the definition (already replaced)
+        # Use word boundary to avoid partial matches
+        modified = re.sub(
+            r"\b" + re.escape(original) + r"\b",
+            placeholder,
+            modified,
+        )
+
+    return modified, name_map
+
+
+class SqlCleaner:
+    """Preprocess raw SQL strings before dialect parsing."""
+
+    @staticmethod
+    def clean(sql: str) -> CleanResult:
+        """Apply all preprocessing steps to raw SQL.
+
+        Steps (in order):
+
+        1. Rewrite ``REPLACE INTO`` → ``INSERT INTO``.
+        2. Rewrite ``SELECT...INTO var FROM`` → ``SELECT...FROM``.
+        3. Strip comments.
+        4. Normalise qualified CTE names.
+        5. Strip DB2 isolation-level clauses.
+        6. Detect malformed ``WITH...AS(...)  AS`` patterns.
+        7. Strip redundant outer parentheses.
+
+        :param sql: Raw SQL string.
+        :type sql: str
+        :returns: Cleaning result with preprocessed SQL (``None`` if
+            effectively empty), replace flag, and CTE name map.
+        :rtype: CleanResult
+        :raises ValueError: If a malformed WITH pattern is detected.
+        """
+        is_replace = False
+        if re.match(r"\s*REPLACE\b", sql, re.IGNORECASE):
+            sql = re.sub(
+                r"\bREPLACE\s+INTO\b",
+                "INSERT INTO",
+                sql,
+                count=1,
+                flags=re.IGNORECASE,
+            )
+            is_replace = True
+
+        # Rewrite SELECT...INTO var1,var2 FROM → SELECT...FROM
+        # so sqlglot doesn't treat variables as tables.
+        sql = re.sub(
+            r"(?i)(\bSELECT\b.+?)\bINTO\b.+?\bFROM\b",
+            r"\1FROM",
+            sql,
+            count=1,
+            flags=re.DOTALL,
+        )
+
+        clean_sql = _strip_comments(sql)
+        if not clean_sql.strip():
+            return CleanResult(sql=None, is_replace=is_replace, cte_name_map={})
+
+        clean_sql, cte_name_map = _normalize_cte_names(clean_sql)
+        clean_sql = re.sub(
+            r"\bwith\s+(ur|cs|rs|rr)\s*$", "", clean_sql, flags=re.IGNORECASE
+        ).strip()
+
+        SqlCleaner._detect_malformed_with(clean_sql)
+
+        clean_sql = _strip_outer_parens(clean_sql)
+        if not clean_sql.strip():
+            return CleanResult(
+                sql=None, is_replace=is_replace, cte_name_map=cte_name_map
+            )
+
+        return CleanResult(
+            sql=clean_sql, is_replace=is_replace, cte_name_map=cte_name_map
+        )
+
+    @staticmethod
+    def _detect_malformed_with(clean_sql: str) -> None:
+        """Raise ``ValueError`` if the SQL contains a malformed WITH pattern.
+
+        Detects ``WITH...AS(...)  AS <keyword>`` or
+        ``WITH...AS(...)  AS <word> <keyword>`` — an extra ``AS`` token
+        after the CTE body that indicates malformed SQL.
+
+        :param clean_sql: Preprocessed SQL string.
+        :type clean_sql: str
+        :raises ValueError: If a malformed WITH pattern is found.
+        """
+        if not re.match(r"\s*WITH\b", clean_sql, re.IGNORECASE):
+            return
+        main_kw = r"(?:SELECT|INSERT|UPDATE|DELETE)"
+        if re.search(
+            r"\)\s+AS\s+" + main_kw + r"\b", clean_sql, re.IGNORECASE
+        ) or re.search(r"\)\s+AS\s+\w+\s+" + main_kw + r"\b", clean_sql, re.IGNORECASE):
+            raise ValueError("This query is wrong")
